src/ai-core/requirements.txt
# Core ML/AI
torch==2.1.0
transformers==4.35.2
sentence-transformers==2.2.2
openai==1.5.0

# Vector DB
chromadb==0.4.18
faiss-cpu==1.7.4

# Text Processing
tiktoken==0.5.2
spacy==3.7.2
nltk==3.8.1

# Data Processing
numpy==1.24.4
pandas==2.1.4
scikit-learn==1.3.2

# Document Parsing
PyYAML==6.0.1
python-docx==1.1.0
PyPDF2==3.0.1

# API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Utils
python-dotenv==1.0.0
structlog==23.2.0
aiofiles==23.2.1